requirements.txt
# requirements.txt (recommended curated set for Python 3.11)
fastapi>=0.95.2,<1.0
uvicorn[standard]>=0.22.0,<1.0
python-dotenv>=1.0.0
httpx>=0.24.1,<1.0
llama-cpp-python>=0.2.23
qdrant-client>=1.15.1
sentence-transformers>=2.2.2,<3.0
langchain>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.0.19
langsmith>=0.4.0
langgraph>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.3.4
tenacity>=8.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.5
